% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RagGrid.R
\name{aggrid}
\alias{aggrid}
\title{Create a HTML widget using the ag-grid library}
\usage{
aggrid(data, options = list(), licenseKey = NULL, width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{data}{a dataobject (either a matrix or a dataframe)}

\item{options}{a list of ag-grid grid options}

\item{licenseKey}{if you wish you to use the enterprise version of ag-grid}

\item{width, height}{Width/Height in pixels (optional, defaults to automatic
sizing)}

\item{elementId}{An id for the widget (a random string by default).}
}
\description{
This function creates a HTML widget to display matrix or a dataframe using ag-grid.
}
