% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formatColumns}
\alias{formatColumns}
\title{Format table columns}
\usage{
formatColumns(table, columns, format)
}
\arguments{
\item{table}{a table object created from \code{\link{aggrid}()}}

\item{columns}{the indices of the columns}

\item{format}{format based on numeral.js (see
\url{http://numeraljs.com/#format});}
}
\description{
Format table columns
}
